archive/demos/01-essentials/01-foundry/02-sdks/foundry-sdk-py/pyproject.toml
[project]
name = "foundry-sdk-py"
version = "0.1.0"
description = "Azure AI Foundry SDK Python Demo"
readme = "readme.md"
requires-python = ">=3.9"
dependencies = [
    "azure-ai-projects",
    "azure-ai-inference",
    "azure-identity",
    "python-dotenv",
    "openai",
]